Add parser as default feature
