+++ Cargo.toml
@@ -9,4 +9,9 @@ documentation = "https://rutgersc.github.io/doc/m3u8_rs/index.html"
 license = "MIT"
 
 [dependencies]
-nom = "5.1.0"
+nom = { version = "5.1.0", optional = true }
+
+[features]
+default = ["parser"]
+parser = ["nom"]
+
